Move git component tests to unit tests
