+++ backend/tests/conftest.py
@@ -14,6 +14,9 @@ import pytest_asyncio
 import ujson
 from fast_depends import Provider
 from fast_depends import dependency_provider as provider
+from infrahub_sdk import Config, InfrahubClient
+from infrahub_sdk.branch import BranchData
+from infrahub_sdk.uuidt import UUIDT
 from neo4j import GraphDatabase
 from neo4j.exceptions import ServiceUnavailable
 from prefect import settings as prefect_settings
@@ -49,6 +52,7 @@ from infrahub.core.schema.relationship_schema import RelationshipSchema
 from infrahub.core.schema.schema_branch import SchemaBranch
 from infrahub.core.utils import delete_all_nodes
 from infrahub.database import InfrahubDatabase
+from infrahub.git import InfrahubRepository
 from infrahub.graphql.manager import registry as graphql_registry
 from infrahub.lock import initialize_lock
 from infrahub.message_bus import InfrahubMessage, InfrahubResponse
@@ -71,6 +75,8 @@ from tests.helpers.constants import (
     PORT_PREFECT,
     PORT_REDIS,
 )
+from tests.helpers.file_repo import FileRepo
+from tests.helpers.test_client import dummy_async_request
 from tests.helpers.utils import get_exposed_port, start_neo4j_container, start_prefect_server_container
 
 ResponseClass = TypeVar("ResponseClass")
@@ -1165,6 +1171,72 @@ def git_repos_source_dir_module_scope(tmp_path_module_scope: Path) -> Path:
     repos_dir = tmp_path_module_scope / "source"
     repos_dir.mkdir()
     return repos_dir
+
+
+@pytest.fixture
+def git_repos_dir(tmp_path: Path) -> Generator[Path, None, None]:
+    repos_dir = tmp_path / "repositories"
+    repos_dir.mkdir()
+    original = config.SETTINGS.git.repositories_directory
+    config.SETTINGS.git.repositories_directory = str(repos_dir)
+    yield repos_dir
+    config.SETTINGS.git.repositories_directory = original
+
+
+@pytest.fixture
+def branch01() -> BranchData:
+    return BranchData(
+        id="6c915158-d8ef-4169-9b00-59f94716b8c3",
+        name="branch01",
+        sync_with_git=False,
+        is_default=False,
+        branched_from="main",
+        has_schema_changes=False,
+    )
+
+
+@pytest.fixture
+def branch02() -> BranchData:
+    return BranchData(
+        id="7708dcea-f7b4-4f5a-b5e9-a0605d4c11ba",
+        name="branch02",
+        sync_with_git=False,
+        is_default=False,
+        branched_from="main",
+        has_schema_changes=False,
+    )
+
+
+@pytest.fixture
+def branch99() -> BranchData:
+    return BranchData(
+        id="2e933717-086c-47cf-8242-21421dd3c2bb",
+        name="branch99",
+        sync_with_git=False,
+        is_default=False,
+        branched_from="main",
+        has_schema_changes=False,
+    )
+
+
+@pytest.fixture
+def git_upstream_repo_01(git_sources_dir: Path) -> dict[str, str | Path]:
+    """Git repository with 4 branches: main, branch01, branch02, and clean-branch.
+    There is a conflict between branch01 and branch02."""
+    name = "infrahub-test-fixture-01"
+    file_repo = FileRepo(name=name, sources_directory=git_sources_dir)
+    return {"name": name, "path": Path(file_repo.path)}
+
+
+@pytest.fixture
+async def git_repo_01(git_upstream_repo_01: dict[str, str | Path], git_repos_dir: Path) -> InfrahubRepository:
+    """Git Repository with git_upstream_repo_01 as remote."""
+    return await InfrahubRepository.new(
+        id=UUIDT.new(),
+        name=git_upstream_repo_01["name"],
+        location=str(git_upstream_repo_01["path"]),
+        client=InfrahubClient(config=Config(requester=dummy_async_request)),
+    )
 
 
 class BusRPCMock(InfrahubMessageBus):
